Struct treerite::Predictor

pub struct Predictor { /* fields omitted */ }

Loader for compiled shared libraries.

There should be at most only one thread calling predict_batch at the same time. As the result, Predictor is Send but not Sync.

Implementations

impl Predictor

pub fn load<P>(
library_path: P,
num_worker_thread: usize
) -> Result<Predictor, TreeRiteError> where
P: AsRef<Path>,

Load the compiled shared libraries from a filesystem location.

Parameters

library_path: location of dynamic shared library (.dll/.so/.dylib)

num_worker_thread: number of worker threads to use; if unspecified, use maximum number of hardware threads.

Example

use treerite::Predictor;
let model = Predictor::load("examples/iris.so", 1).unwrap();

pub fn predict_batch<F, O>(
&self,
dmatrix: &DMatrix<F>,
verbose: bool,
pred_margin: bool
) -> Result<Array2<O>, TreeRiteError> where
O: Float + FloatInfo,
F: Float + FloatInfo,

Perform batch prediction with a 2D sparse data matrix. Worker threads will internally divide up work for batch prediction.

Example

use treerite::{Predictor, DMatrix};
use ndarray::Array2;
let model = Predictor::load("examples/iris.so", 1).unwrap();
let feat: Vec<f64> = vec![7.7, 2.8, 6.7, 2.];
let pred: Array2<f64> = model.predict_batch(&DMatrix::from_slice(&feat).unwrap(), false, false).unwrap();

Note

The output array should be explicitly typed, i.e. f64 or f32.

pub fn leaf_output_type(&self) -> Result<DataType, TreeRiteError>

Get the output type of the leaves.

pub fn num_class(&self) -> Result<u64, TreeRiteError>

Get how many classes this model output

pub fn result_size<F>(&self, batch: &DMatrix<F>) -> Result<u64, TreeRiteError>

Get the output result size (num instances x num classes) given a DMatrix

pub fn num_feature(&self) -> Result<u64, TreeRiteError>

Get the number of features this model required

pub fn global_bias(&self) -> Result<f32, TreeRiteError>

Get the global bias of the model

pub fn sigmod_alpha(&self) -> Result<f32, TreeRiteError>

Get the sigmoid alpha of the model

pub fn pred_transform(&self) -> Result<String, TreeRiteError>

pub fn threshold_type(&self) -> Result<String, TreeRiteError>